tnk_quest_back/requirements.txt
Django==4.1.12
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==3.10.0
python_http_client
cryptography
pyyaml
uwsgi